refactor: update compression member configuration to use renamed and standardized asset images
