+++ frontend/src/modules/compressionMember/configs/strutsWeldedConfig.js
@@ -1,8 +1,8 @@
-import ANGLES from "../../../assets/CompressionMember/wA.png";
-import BACK_TO_BACK_ANGLES from "../../../assets/CompressionMember/wBBA.png";
-import STAR_ANGLES from "../../../assets/CompressionMember/wSA.png";
-import CHANNELS from "../../../assets/CompressionMember/wC.png";
-import BACK_TO_BACK_CHANNELS from "../../../assets/CompressionMember/wBBC.png";
+import ANGLES from "../../../assets/CompressionMember/bA.png";
+import BACK_TO_BACK_ANGLES from "../../../assets/CompressionMember/back_back_same_side_angles.png";
+import STAR_ANGLES from "../../../assets/CompressionMember/bBBA.png";
+import CHANNELS from "../../../assets/CompressionMember/bC.png";
+import BACK_TO_BACK_CHANNELS from "../../../assets/CompressionMember/bBBC.png";
 import ErrorImg from "../../../assets/notSelected.png";
 import FIXED_FIXED from "../../../assets/CompressionMember/RRRRstrut.png";
 import FIXED_HINGED from "../../../assets/CompressionMember/RFRFstrut.png";
